Python code
The python code that scrapes the SEC website and finds companies 10k documents since 2011 to 2023 if possible. It then creates a master list of the companies and in it has the creation of the ID which is then used as the directory names for each of the companies. Then using the keywords in the Keywords.csv file it will create a csv file of a matrix of the frequency of each keyword for each year.